old version files removal, WiFiConnectionManager
