--- WiFiQR/Old Unused Version/Models/WiFiCell.swift
@@ -1,21 +1,0 @@
-//
-//  WiFiCell.swift
-//  WIFIQR
-//
-//  Created by riccardo silvi on 27/12/17.
-//  Copyright © 2017 riccardo silvi. All rights reserved.
-//
-
-import UIKit
-
-class WiFiCell: UITableViewCell {
-
-    @IBOutlet weak var qrcodeImageView: UIImageView!
-
-    @IBOutlet weak var nomeReteWiFiLabel: UILabel!
-    
-    @IBOutlet weak var visibilitàReteWiFiLabel: UILabel!
-    
-    @IBOutlet weak var crittazionePasswordLabel: UILabel!
-    
-}
